+++ tests/Functional/DoctrineOrm/Entities/Total.php
@@ -1,0 +1,23 @@
+<?php
+
+/*
+ * This file is part of the FiveLab Ruler package
+ *
+ * (c) FiveLab
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code
+ */
+
+declare(strict_types = 1);
+
+namespace FiveLab\Component\Ruler\Tests\Functional\DoctrineOrm\Entities;
+
+use Doctrine\ORM\Mapping as ORM;
+
+#[ORM\Embeddable]
+class Total
+{
+    #[ORM\Embedded(class: Money::class, columnPrefix: 'money_')]
+    private Money $money;
+}